Layout working for most Geometry aside from the various edge cases we need to account for like :
Corners of a rounded rectangle. 
Starting draw at a specific x,y when the left and top are not used on path and poly objects.
Enforcing aspect ratio for circle.
ETC..

The default values for most geometry properties have been updated to allow layout to do it’s job. For example if no radius is specified on a circle layout still needs a minimum value other than 0 to work off of. Because of this hasValideProperties has been added to test the specific Geometry for the minimum combination of properties required to render the object. In the case of circle in order to render we require either a radius or a layout.

Root geometry will layout to the current document if no graphicsTarget(s) specified. This was required for layout to properly work in a skin and is temporary until I find a more appropriate solution.

The rules still have to be decided/completed on how to properly handles these edge cases and defaults.

Still lots to do/add/test